pyproject.toml
[tool.poetry]
name = "google-drive-docs-api-drf"
version = "0.1.0"
description = ""
authors = ["vova2092 <vovai2092@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
pydrive2 = "^1.19.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
flake8 = "^7.0.0"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
